5.1.4 Рефакторинг модуля с утилитами
