--- src/utils.js
@@ -1,76 +1,0 @@
-export const RenderPosition = {
-  AFTER_BEGIN: `afterbegin`,
-  BEFOR_END: `beforeend`,
-};
-
-export const renderTemplate = (container, template, place = RenderPosition.BEFOR_END) => {
-  container.insertAdjacentHTML(place, template);
-};
-
-export const render = (container, element, place = RenderPosition.BEFOR_END) => {
-  switch (place) {
-    case RenderPosition.AFTER_BEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOR_END:
-      container.append(element);
-      break;
-  }
-};
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-// Функция генерации случайного целого числа из диапазона
-// Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
-export const getRandomInteger = (a = 1, b = 0) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-export const getRandomArrayItems = (array) => {
-  const randomIndex = getRandomInteger(array.length - 1);
-
-  return array[randomIndex];
-};
-
-const getCurrentDate = () => {
-  const currentDate = new Date();
-  currentDate.setHours(23, 59, 59, 999);
-
-  return new Date(currentDate);
-};
-
-export const isExpiredTask = (dueDate) => {
-  if (dueDate === null) {
-    return false;
-  }
-
-  const currentDate = getCurrentDate();
-
-  return currentDate.getTime() > dueDate.getTime();
-};
-
-export const isExpiredTaskToday = (dueDate) => {
-  if (dueDate === null) {
-    return false;
-  }
-
-  const currentDate = getCurrentDate();
-
-  return currentDate.getTime() === dueDate.getTime();
-};
-
-export const isRepeatingTask = (repeating) => {
-  return Object.values(repeating).some(Boolean);
-};
-
-export const humanizeTaskDate = (date) => {
-  return date.toLocaleString(`en-GB`, {day: `numeric`, month: `long`});
-};
