+++ src/utils/common.js
@@ -1,0 +1,14 @@
+// Функция генерации случайного целого числа из диапазона
+// Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
+export const getRandomInteger = (a = 1, b = 0) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+
+  return Math.floor(lower + Math.random() * (upper - lower + 1));
+};
+
+export const getRandomArrayItems = (array) => {
+  const randomIndex = getRandomInteger(array.length - 1);
+
+  return array[randomIndex];
+};
