+++ src/utils/task.js
@@ -1,0 +1,34 @@
+const getCurrentDate = () => {
+  const currentDate = new Date();
+  currentDate.setHours(23, 59, 59, 999);
+
+  return new Date(currentDate);
+};
+
+export const isExpiredTask = (dueDate) => {
+  if (dueDate === null) {
+    return false;
+  }
+
+  const currentDate = getCurrentDate();
+
+  return currentDate.getTime() > dueDate.getTime();
+};
+
+export const isExpiredTaskToday = (dueDate) => {
+  if (dueDate === null) {
+    return false;
+  }
+
+  const currentDate = getCurrentDate();
+
+  return currentDate.getTime() === dueDate.getTime();
+};
+
+export const isRepeatingTask = (repeating) => {
+  return Object.values(repeating).some(Boolean);
+};
+
+export const humanizeTaskDate = (date) => {
+  return date.toLocaleString(`en-GB`, {day: `numeric`, month: `long`});
+};
